% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion_header.R
\name{notion_header}
\alias{notion_header}
\title{Function to generate a request header for notion.so}
\usage{
notion_header(version = "2021-05-13")
}
\arguments{
\item{version}{Notion version used in a request}
}
\description{
Function to generate a request header for notion.so
}
\details{
The Notion API is versioned and named for the date the version
is released. The current latest version is 2021-05-13.
please check: \url{https://developers.notion.com/reference/versioning}
}
\examples{
notion_header()

}
\seealso{
"set_notin_token"
}
