% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion_search.R
\name{notion_search}
\alias{notion_search}
\title{Notion's Search API}
\usage{
notion_search(
  query = NULL,
  sort = NULL,
  filter = NULL,
  start_cursor = NULL,
  page_size = NULL,
  api_rate_limited = 1/3,
  ...
)
}
\arguments{
\item{query}{a length 1 character vector. When supplied, limits which pages
are returned by comparing the query to the page title.}

\item{sort}{a notion sort object generated by \code{\link{obj_sort}} function.
When supplied, sorts the results based on the provided criteria.}

\item{filter}{a notion filter object generated by \code{\link{obj_filter}} function.
When supplied, filters the results based on the provided criteria.}

\item{start_cursor}{a length 1 character vector. If supplied, this endpoint
will return a page of results starting after the cursor provided.
If not supplied, this endpoint will return the first page of results.}

\item{page_size}{a length 1 integer vector. The number of items from the full
list desired in the response. Maximum: 100}

\item{api_rate_limited}{The rate limit for incoming requests is an average of
3 requests per second. Some bursts beyond the average rate are allowed.}
}
\description{
Wrapper function for Notion's Search API'
}
\details{
please check: \url{https://developers.notion.com/reference/post-search}
}
\examples{
notion_search()

}
