% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{rni_add_page}
\alias{rni_add_page}
\title{Insert a page into a database or page}
\usage{
rni_add_page(parent, properties, ...)
}
\arguments{
\item{parent}{A list with type (\code{database_id} or \code{page_id}) and the corresponding id}

\item{properties}{A list of properties}

\item{...}{}
}
\value{
Nothing
}
\description{
Insert a page into a database or page
}
