% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_annotate_text}
\alias{rni_annotate_text}
\title{Format notion text}
\usage{
rni_annotate_text(text)
}
\arguments{
\item{text}{A notion text object}
}
\value{
A formated string
}
\description{
Format notion text
}
