% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_date}
\alias{rni_date}
\title{Transform date properties}
\usage{
rni_date(block)
}
\arguments{
\item{block}{A notion block object}
}
\value{
A lubridate interval
}
\description{
Transform date properties
}
\keyword{internal}
