% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rni_get_database}
\alias{rni_get_database}
\title{Get all elements of a database}
\usage{
rni_get_database(id, cursor = NULL, ...)
}
\arguments{
\item{id}{The database id}

\item{cursor}{An optional cursor for multipage requests}

\item{sort_params}{A list of sort parameters}
}
\value{
A list of properties
}
\description{
Get all elements of a database
}
