% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{rni_get_page_properties}
\alias{rni_get_page_properties}
\title{Get propoerties from a page}
\usage{
rni_get_page_properties(id)
}
\arguments{
\item{id}{A notion page id}
}
\value{
A list
}
\description{
Get propoerties from a page
}
