% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rni_map_multi_select_list}
\alias{rni_map_multi_select_list}
\title{Make a list for multi-select properties}
\usage{
rni_map_multi_select_list(category)
}
\arguments{
\item{category}{A string}
}
\value{
A list
}
\description{
Make a list for multi-select properties
}
