% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{rni_page_properties}
\alias{rni_page_properties}
\title{Process page properties}
\usage{
rni_page_properties(page)
}
\arguments{
\item{page}{A notion page object}
}
\value{
A string
}
\description{
Process page properties
}
\keyword{internal}
