% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_parse_block}
\alias{rni_parse_block}
\title{Parse notion blocks}
\usage{
rni_parse_block(block, render_level)
}
\arguments{
\item{block}{A notion block object}
}
\value{

}
\description{
Parse notion blocks
}
