% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rni_properties_tibble}
\alias{rni_properties_tibble}
\title{Format a notion database as tibble}
\usage{
rni_properties_tibble(db)
}
\arguments{
\item{db}{A list of notion pages}
}
\value{
A tibble
}
\description{
Format a notion database as tibble
}
