% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_render_blocks}
\alias{rni_render_blocks}
\title{Recursively render a block}
\usage{
rni_render_blocks(id, render_level = 0)
}
\arguments{
\item{id}{The block id string}
}
\value{
A markdown string
}
\description{
Recursively render a block
}
