% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_rich_text}
\alias{rni_rich_text}
\title{Apply text formatting for rich text properties}
\usage{
rni_rich_text(block)
}
\arguments{
\item{block}{A notion block object}
}
\value{
A vector of strings
}
\description{
Apply text formatting for rich text properties
}
\keyword{internal}
