% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_text}
\alias{rni_text}
\title{Apply text formatting}
\usage{
rni_text(block)
}
\arguments{
\item{block}{A notion block object}
}
\value{
A vector of strings
}
\description{
Apply text formatting
}
\keyword{internal}
