% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rni_to_md}
\alias{rni_to_md}
\title{Format notion block into markdown}
\usage{
rni_to_md(block)
}
\arguments{
\item{block}{A notion block object}
}
\value{
A string
}
\description{
Format notion block into markdown
}
